remove dependence of Emulator.h in Wave.h
