+++ shared/race-emulator/include/race-emulator/Emulator.h
@@ -4,6 +4,7 @@
 #pragma once
 #include "Arch.h"
 #include "Profiler.h"
+#include "Util.h"
 #include "Wave.h"
 #include "Workgroup.h"
 #include <cassert>
