+++ shared/race-emulator/include/race-emulator/Workgroup.h
@@ -6,49 +6,13 @@
 #include "EmulatorException.h"
 #include "IntervalSet.h"
 #include "LDS.h"
+#include "Types.h"
 #include <cassert>
 #include <cstdint>
 #include <string>
 #include <vector>
 
 namespace raceemulator {
-
-/// Identifies a wave (SIMD execution unit) within a workgroup. Wave 0 runs
-/// lanes [0, waveSize), wave 1 runs [waveSize, 2*waveSize), and so on.
-/// Strongly typed to prevent accidental mixing with lane indices, register
-/// indices, or event IDs.
-struct WaveId {
-  int value;
-  bool operator==(WaveId o) const { return value == o.value; }
-  bool operator!=(WaveId o) const { return value != o.value; }
-  bool operator<(WaveId o) const { return value < o.value; }
-};
-
-/// Number of lanes per wave (32 for RDNA, 64 for CDNA).
-/// Implicitly convertible to int for use in arithmetic and loop bounds.
-struct WaveSize {
-  int value;
-  operator int() const { return value; }
-};
-
-/// Identifies a memory event within a workgroup. Each memory instruction
-/// (LDS read/write, global load/store) creates an event via
-/// Workgroup::allocateEventId(). The event ID is used to track the event
-/// through its lifecycle: registration, s_waitcnt completion, and barrier
-/// retirement. Strongly typed to prevent accidental mixing with wave IDs,
-/// register indices, or byte addresses.
-struct EventId {
-  int value;
-  bool operator==(EventId o) const { return value == o.value; }
-  bool operator!=(EventId o) const { return value != o.value; }
-  bool operator<(EventId o) const { return value < o.value; }
-};
-
-/// Status of a memory event in the race detection lifecycle.
-enum class EventStatus {
-  ACTIVE,       // Pending. Unsafe for everyone.
-  WAVE_COMPLETE // s_waitcnt passed. Safe for owning wave, unsafe for others.
-};
 
 /// Workgroup-scoped state: owns LDS memory, the event registry, and race
 /// detection logic. Waves hold a reference to a Workgroup and use it for
